Extract duplicated fetch/undici utils into shared file
Both the full mode (SentryNodeFetchInstrumentation) and light mode
(nativeNodeFetchIntegration) duplicated identical logic for trace
propagation header injection, breadcrumb creation, and URL construction.
Extract these into a shared outgoingFetchRequest.ts utility.
