+++ packages/node-core/src/integrations/node-fetch/SentryNodeFetchInstrumentation.ts
@@ -2,28 +2,15 @@ import { context } from '@opentelemetry/api';
 import { isTracingSuppressed } from '@opentelemetry/core';
 import type { InstrumentationConfig } from '@opentelemetry/instrumentation';
 import { InstrumentationBase } from '@opentelemetry/instrumentation';
-import type { SanitizedRequestData } from '@sentry/core';
-import {
-  addBreadcrumb,
-  getBreadcrumbLogLevelFromHttpStatusCode,
-  getClient,
-  getSanitizedUrlString,
-  getTraceData,
-  LRUMap,
-  parseUrl,
-  SDK_VERSION,
-  shouldPropagateTraceForUrl,
-} from '@sentry/core';
+import { LRUMap, SDK_VERSION } from '@sentry/core';
 import * as diagch from 'diagnostics_channel';
 import { NODE_MAJOR, NODE_MINOR } from '../../nodeVersion';
-import { mergeBaggageHeaders } from '../../utils/baggage';
+import {
+  addFetchRequestBreadcrumb,
+  addTracePropagationHeadersToFetchRequest,
+  getAbsoluteUrl,
+} from '../../utils/outgoingFetchRequest';
 import type { UndiciRequest, UndiciResponse } from './types';
-
-const SENTRY_TRACE_HEADER = 'sentry-trace';
-const SENTRY_BAGGAGE_HEADER = 'baggage';
-
-// For baggage, we make sure to merge this into a possibly existing header
-const BAGGAGE_HEADER_REGEX = /baggage: (.*)\r\n/;
 
 export type SentryNodeFetchInstrumentationOptions = InstrumentationConfig & {
   /**
@@ -114,7 +101,6 @@ export class SentryNodeFetchInstrumentation extends InstrumentationBase<SentryNo
    * This method is called when a request is created.
    * You can still mutate the request here before it is sent.
    */
-  // eslint-disable-next-line complexity
   private _onRequestCreated({ request }: { request: UndiciRequest }): void {
     const config = this.getConfig();
     const enabled = config.enabled !== false;
@@ -132,70 +118,7 @@ export class SentryNodeFetchInstrumentation extends InstrumentationBase<SentryNo
       return;
     }
 
-    const url = getAbsoluteUrl(request.origin, request.path);
-
-    // Manually add the trace headers, if it applies
-    // Note: We do not use `propagation.inject()` here, because our propagator relies on an active span
-    // Which we do not have in this case
-    // The propagator _may_ overwrite this, but this should be fine as it is the same data
-    const { tracePropagationTargets, propagateTraceparent } = getClient()?.getOptions() || {};
-    const addedHeaders = shouldPropagateTraceForUrl(url, tracePropagationTargets, this._propagationDecisionMap)
-      ? getTraceData({ propagateTraceparent })
-      : undefined;
-
-    if (!addedHeaders) {
-      return;
-    }
-
-    const { 'sentry-trace': sentryTrace, baggage, traceparent } = addedHeaders;
-
-    // We do not want to overwrite existing headers here
-    // If the core UndiciInstrumentation is registered, it will already have set the headers
-    // We do not want to add any then
-    if (Array.isArray(request.headers)) {
-      const requestHeaders = request.headers;
-
-      // We do not want to overwrite existing header here, if it was already set
-      if (sentryTrace && !requestHeaders.includes(SENTRY_TRACE_HEADER)) {
-        requestHeaders.push(SENTRY_TRACE_HEADER, sentryTrace);
-      }
-
-      if (traceparent && !requestHeaders.includes('traceparent')) {
-        requestHeaders.push('traceparent', traceparent);
-      }
-
-      // For baggage, we make sure to merge this into a possibly existing header
-      const existingBaggagePos = requestHeaders.findIndex(header => header === SENTRY_BAGGAGE_HEADER);
-      if (baggage && existingBaggagePos === -1) {
-        requestHeaders.push(SENTRY_BAGGAGE_HEADER, baggage);
-      } else if (baggage) {
-        const existingBaggage = requestHeaders[existingBaggagePos + 1];
-        const merged = mergeBaggageHeaders(existingBaggage, baggage);
-        if (merged) {
-          requestHeaders[existingBaggagePos + 1] = merged;
-        }
-      }
-    } else {
-      const requestHeaders = request.headers;
-      // We do not want to overwrite existing header here, if it was already set
-      if (sentryTrace && !requestHeaders.includes(`${SENTRY_TRACE_HEADER}:`)) {
-        request.headers += `${SENTRY_TRACE_HEADER}: ${sentryTrace}\r\n`;
-      }
-
-      if (traceparent && !requestHeaders.includes('traceparent:')) {
-        request.headers += `traceparent: ${traceparent}\r\n`;
-      }
-
-      const existingBaggage = request.headers.match(BAGGAGE_HEADER_REGEX)?.[1];
-      if (baggage && !existingBaggage) {
-        request.headers += `${SENTRY_BAGGAGE_HEADER}: ${baggage}\r\n`;
-      } else if (baggage) {
-        const merged = mergeBaggageHeaders(existingBaggage, baggage);
-        if (merged) {
-          request.headers = request.headers.replace(BAGGAGE_HEADER_REGEX, `baggage: ${merged}\r\n`);
-        }
-      }
-    }
+    addTracePropagationHeadersToFetchRequest(request, this._propagationDecisionMap);
   }
 
   /**
@@ -215,7 +138,7 @@ export class SentryNodeFetchInstrumentation extends InstrumentationBase<SentryNo
     const shouldIgnore = this._ignoreOutgoingRequestsMap.get(request);
 
     if (breadCrumbsEnabled && !shouldIgnore) {
-      addRequestBreadcrumb(request, response);
+      addFetchRequestBreadcrumb(request, response);
     }
   }
 
@@ -261,73 +184,5 @@ export class SentryNodeFetchInstrumentation extends InstrumentationBase<SentryNo
     }
 
     return ignoreOutgoingRequests(url);
-  }
-}
-
-/** Add a breadcrumb for outgoing requests. */
-function addRequestBreadcrumb(request: UndiciRequest, response: UndiciResponse): void {
-  const data = getBreadcrumbData(request);
-
-  const statusCode = response.statusCode;
-  const level = getBreadcrumbLogLevelFromHttpStatusCode(statusCode);
-
-  addBreadcrumb(
-    {
-      category: 'http',
-      data: {
-        status_code: statusCode,
-        ...data,
-      },
-      type: 'http',
-      level,
-    },
-    {
-      event: 'response',
-      request,
-      response,
-    },
-  );
-}
-
-function getBreadcrumbData(request: UndiciRequest): Partial<SanitizedRequestData> {
-  try {
-    const url = getAbsoluteUrl(request.origin, request.path);
-    const parsedUrl = parseUrl(url);
-
-    const data: Partial<SanitizedRequestData> = {
-      url: getSanitizedUrlString(parsedUrl),
-      'http.method': request.method || 'GET',
-    };
-
-    if (parsedUrl.search) {
-      data['http.query'] = parsedUrl.search;
-    }
-    if (parsedUrl.hash) {
-      data['http.fragment'] = parsedUrl.hash;
-    }
-
-    return data;
-  } catch {
-    return {};
-  }
-}
-
-function getAbsoluteUrl(origin: string, path: string = '/'): string {
-  try {
-    const url = new URL(path, origin);
-    return url.toString();
-  } catch {
-    // fallback: Construct it on our own
-    const url = `${origin}`;
-
-    if (url.endsWith('/') && path.startsWith('/')) {
-      return `${url}${path.slice(1)}`;
-    }
-
-    if (!url.endsWith('/') && !path.startsWith('/')) {
-      return `${url}/${path.slice(1)}`;
-    }
-
-    return `${url}${path}`;
   }
 }
